set rpath on Linux and macOS builds
so qsv can find the python shared library in the present working directory, which makes it easier to distribute on Linux and Mac.

Windows doesn't need this as it automatically picks up the python DLL if its in the same directory as qsv.

also went ahead and turned on CPU optimization for Intel macOS builds.
[skip ci]