Fixes qsv#444 - New Multi-Column feature for `apply` command
With this CL, we add a new feature of applying operations to
multiple columns simultaneously,

- qsv apply operations trim,upper name,surname ~/userdetails.csv
- qsv apply operations trim,upper 1,2 ~/userdetails.csv

In addition of selecting multiple columns, we support renaming them.
However, the no of columns being selected should be equal to number
of columns being renamed.

So this is valid:

- qsv apply operations trim,upper name,surname --rename uname,usurname
  ~/userdetails.csv

But this is invalid:

- qsv apply operations trim,upper name,surname --rename uname ~/userdetails.csv

We do not support multi column operations for apply subcommands `dynfmt`
and `geocode`. As it logically does not make sense.