diff: Implement new command `diff`
This implements the new command `diff` for diffing two CSVs.
Usage is as follows:

Usage:
    qsv diff [options] [<input-left>] [<input-right>]
    qsv diff --help

Common options:
    -h, --help                  Display this message
    -o, --output <file>         Write output to <file> instead of stdout.
    --no-headers-left           When set, the first row will be considered as part of
                                the left CSV to diff. (When not set, the
                                first row is the header row and will be skipped during
                                the diff. It will always appear in the output.)
    --no-headers-right          When set, the first row will be considered as part of
                                the right CSV to diff. (When not set, the
                                first row is the header row and will be skipped during
                                the diff. It will always appear in the output.)
    --delimiter-left <arg>      The field delimiter for reading CSV data on the left.
                                Must be a single character. (default: ,)
    --delimiter-right <arg>     The field delimiter for reading CSV data on the right.
                                Must be a single character. (default: ,)
    --primary-key-idx <arg...>  The column indices that uniquely identify a record
                                as a comma separated list of indices, e.g. 0,1,2.
                                (default: 0)

----------------------------------------------------------------------

Examples:
Diff csv_left.csv with csv_right.csv with default options and write
the result to stdout:

`qsv diff path/to/csv_left.csv path/to/csv_right.csv`

Diff csv_left.csv with csv_right.csv and treat the first and third
columns as unique keys for identifying a record. The result is written
to the new file diffresult.csv:

`qsv diff --primary-key-idx 0,2 -o path/to/diffresult.csv \
path/to/csv_left.csv path/to/csv_right.csv`

The output of the diff command can look like the following:

diffresult,GlobalRank,TldRank,Domain,TLD,RefSubNets,RefIPs,IDN_Domain,IDN_TLD,PrevGlobalRank,PrevTldRank,PrevRefSubNets,PrevRefIPs
-,4,4,twitter.com,com,436953,2075089,twitter.com,com,4,4,436735,2077023
+,4,4,joinmastodon.org,org,436953,2075089,joinmastodon.org,org,4,4,436735,2077023
-,18,14,maps.google.com,com,203146,708101,maps.google.com,com,18,14,203197,708631
-,1,1,google.com,com,492918,2427561,google.com,com,1,1,492881,2429667
-,23,17,docs.google.com,com,184531,539376,docs.google.com,com,23,17,184740,540129
+,19,1,goo.gl,gl,200037,649215,goo.gl,gl,19,1,200069,649539
-,20,15,plus.google.com,com,198766,712276,plus.google.com,com,20,15,198784,713207
-,15,11,play.google.com,com,216317,611475,play.google.com,com,15,11,216267,612046

The diff result can be seen in the first column:
- "-" means that the record was deleted in the file on the right
(can be seen after the first two records in the example output)
- "+" means that the record was added in the file on the right
- consecutive "-", "+", where the key columns are equal means the record
was modified in some columns that were not key columns
(this can be seen in the first two rows of the example output)

The following is an svg recording of a terminal, in which this new
command `diff` is executed and `time`d on the majestic million csv with
some artificial changes introduced.
The CSVs have 1,000,000 rows x 12 columns:
https://gist.githubusercontent.com/janriemer/0e86a236a7174af6faed1b981815a5ea/raw/2ab235a2bdd5798f2348d42da48fdda276fb5f0f/qsv_diff_recording.svg

------------------------------------------------------------------------

New dependency `csv-diff`:
The actual operation of diffing two CSVs is provided by the external
crate `csv-diff`, so this uses this new dependency:
https://crates.io/crates/csv-diff
(see also 90a7035, where this dependency is introduced)

`csv-diff` uses a default feature where a rayon thread pool is used.
As `qsv` doesn't use a global thread pool, this feature might be
unnecessarily activated in `csv-diff`. In future iterations this default
feature should be considered to be disabled, but for now we leave it as
is to make things simple.

Please keep in mind that `csv-diff` is in an early stage of development,
so the new `diff` command should be considered experimental for now.

------------------------------------------------------------------------

Current limitations:
- if both CSVs have headers, they must not be in a different ordering
=> see also these issues in `csv-diff`:
    - https://gitlab.com/janriemer/csv-diff/-/issues/6
    - https://gitlab.com/janriemer/csv-diff/-/issues/3

------------------------------------------------------------------------

Closes #547