Update csvs_convert.
The main change is adding --dump option so that you can produce a
dump file that will work with `psql` for postgres and `sqlite3` for
sqlite.
This uses the native import functions '\copy` for postgres
(which allows remote copy) and `.import` for sqlite, which is the
fastest way to insert into sqlite.
This dumpfile is also useful to just
inspect what the command will produce if you run it directly against a
database. This is to addresses partially
https://github.com/jqnatividad/qsv/issues/687 as it is the representation
that most replicates what the command will do.

Other changed include making sure strings that start with 0 (but are not
0) are treated as strings not numbers which should resolve
https://github.com/jqnatividad/qsv/issues/688