`diff`: rearrange test for sorting by column
The diffresult was difficult to see in the test, because it was pulled
from a file. It is now hard coded in the actual test code for better
maintainability.

Note that we can only do this, because the diffresult is pretty small
in terms of lines (10 lines). The actual files to diff are pretty large,
so it _does_ make sense to keep them external in a dedicated file.