`diff`: rename --primary-key-idx -> --key
This is more succinct and more in alignment with other tools, such as
Python's `csv-diff` tool:
https://pypi.org/project/csv-diff/

At the moment only column __indices__ are supported, but column/header
__names__ might also be possible in the future.

Note, that we can later mix and match them easily, e.g.
--key 0,1,id,\"2\" <-- key that consists of first two columns and
columns with name "id" and "2", so there is no need for dedicated
`--key-idx` and `--key-headers` flags.

BREAKING: Use `--key` or `-k` instead of `--primary-key-idx`.

Part of #714