`diff`: Implement option to sort by columns
This implements the option for `diff` to sort by multiple columns.

Example (where the first column is the key):

left.csv
```
id1,a,c
id2,1,a
id3,2,6
id4,b,_
```
right.csv
```
id1,a,a
id7,1,b
id8,2,5
id4,a,_
```

```
qsv diff --sort-columns 1,2 left.csv right.csv
```

This will sort the resulting diff by column with index 1 (and 2, if the
field of two diff records compare as equal). The comparison is done
based on `&[u8]` (which is the content of a field as a byte slice).

Based on the above example, the sorted diffresult looks like this:
```
-,id2,1,a
+,id7,1,b
+,id8,2,5
-,id3,2,6
-,id1,a,c
+,id1,a,a
-,id4,b,_
+,id4,a,_
```
Note that modified records (-+ records with the same key) __are always
kept together__, which makes the result not fully sorted between those
pairs. This is a current limitation of the underlying library
`csv-diff`. This limitation _might_ be resolved in the future
(also see https://gitlab.com/janriemer/csv-diff/-/issues/14). This
limitation can also be seen as an advantage, though, because one can
more easily see the related record of the one modified row.

Currently, comparison against records of type `Modify` are done against
the `delete` record of the `Modify` record. If this comparison is equal,
it will fall back to comparing it with the `add` record of the `Modify`
record
(also see https://gitlab.com/janriemer/csv-diff/-/blob/27b80a9f42183667b431df2b91dacfa4b8f457d1/src/diff_result.rs#L1166-1187).

Note that sorting by column can fail, if one provides an index that is
out of bounds.

Closes #714