`stats`: refactor stats bincode caching
Currently, stats automatically caches stats and serializes it into a bincode file.

This is done so other qsv commands (currently, only `schema` and `tojsonl`) can deserialize the precomputed stats and skip recomputing stats if they are still valid.

As `schema` and `tojsonl` are not as often used, we should only generate the bincode file when requested.

Other changes:
- bincode file is now snappy compressed
- changed stats-binout to a bool flag from an Option<String> parm. We always write stats bincode cache to <FILESTEM>.stats.csv.bin.sz
- removed unneeded bufwriter as snappy encoding has built-in buffering
- expanded cache section in usage text
-