`frequency`: add `--unq-limit` option
This is particularly useful when compiling a frequency table for a column with all unique values (e.g. an ID or Sequence column) as without it, and if --limit is set to zero, the frequency table will be return all the unique values.

There are times when --limit is set to zero when you want to compile all the unique values for a column, but you want to exclude ID/sequence columns from the comprehensive frequency table.