`publish`: add "portable" workflow with CPU features not enabled
when you get an "illegal instruction" error, most likely its because the qsv binary is compiled with CPU features not available on the current system.

CPU features that often increase performance, but are not universally available on all CPUs even for the same target platform (especially, on Virtual Machines)

The "portable" binaries do not have these CPU features enabled.

[skip ci]