`publish`: refactor portable binaries so self-update works
self-update crate requires that the zip archive be named a certain way, and having the `-portable` suffix on the archive prevents that from working, and also doubles the number of zips in each release.

Just create a subvariant binary with the p suffix after qsv to indicate portable binaries - i.e.  qsvp, qsvplite, qsvpdp

[skip ci]