`diff`: add flag `--drop-equal-fields`
This implements a new flag for the command `diff`. When activated, it
drops the values of fields that are equal within a row of type
`Modified` and replaces them with the empty string (an empty byte slice
to be precise). For now, the value for replacing equal values is not
configurable, but should be trivial to add in the future.

Note that key field values are _not_ dropped and always appear in the
output.

Example:
csv_left.csv    col1,col2,col3
                1,foo,bar

csv_right.csv   col1,col2,col3
                1,foo,baz

qsv diff --drop-equal-fields csv_left.csv csv_right.csv

Output:         diffresult,col1,col2,col3
                -,1,,bar
                +,1,,baz

See #2000