feat: `template` add `--outsubdir-size` option
when rendering files to a directory, have a simple mechanism to distribute files across subdirectories of --outsubdir-size.

This makes it easier to handle and explore generated files, especially for large input CSVs.

Also ensures the filesystem remains performant as it severely slows down when there are a large number of files in a directory